Created new abstraction type StaticFunction.
Modified Delegate to use the new type Staticfunction and made the RAII handle type work.

DelegateMulticast now can use directly the StaticFunction type. It is cleaner than using a vector of Delegate type.

Removed the bindObject function so that  all use the bind function and passes the function pointer as NTTP.